Add files via upload
listen command changed to allow mentioning a channel to make it the announcement channel. If none given, it is still the channel the command was given in.
add command now allows mentioning a channel after the stream name to make that stream announce in the mentioned channel, rather than the default listen channel.
addmult also now allows mentioning a channel to set a channel override for all the channels given. It can be anywhere in the command.
addmult now strips a trailing comma from channel names if present, to allow for copy/pasting the output of the list command without modification.
remove and removemult commands now deletes/stops editing current announcement for the deleted stream.
Fixed an issue where streams were marked as offline before the proper wait time.
BACKEND IMPROVEMENTS
Added resolvechannel to APIContext, which determines what channel to send a message to based on server, channel overrides, and eventually global settings.  
# of updates before declaring a stream offline moved to variable, as it's used in multiple places now.
removemsg in APIContext now accepts a list of servers to remove from
Bot no longer attempts to split message until all checks are true - very minor optimization.