Add files via upload
Bot now checks it has send_messages permission in the channel when told to listen. If not, it'll attempt to send a message about the issue in the channel the command was sent from.
removemult now allows a trailing ',' at the end of names, similar to addmult. Again, this makes it easier to use the output of list in the command.
Updated the API class template for the new functions.
Fixed issue with Twitch API returning '' for the game id. Now handled same as 0: 'No game set'