Prep for Discord.py 2.4.0
Make initial required changes:
Explicitly declare Intents.
Fix issues with loop changes and create_task. DB_Client inherits from Client.
myconn is now set as an attribute to client to be used in the setup_hook event that is run at bot start to setup the tasks.
Code moved around to support this - client initialized further down in the code.
Fix for removal of discord.Embed.empty - now uses None.

As of this, bot runs and contexts appear to be working but needs much more testing.

Remove old code to fix aiohttp error in Python 3.7-3.7.3. That version is no longer supported.